refactor: remove unnecessary indirection in internal imports
