+++ packages/core/src/internal/execution/ExecutionGraph.ts
@@ -1,4 +1,4 @@
-import { Graph } from "../../internal/graph/Graph";
+import { Graph } from "../graph/Graph";
 import { ExecutionVertex } from "../types/executionGraph";
 
 export class ExecutionGraph extends Graph<ExecutionVertex> {}
