+++ packages/core/src/internal/process/transform/reduceDeploymentGraphByEliminatingVirtualVertexes.ts
@@ -1,5 +1,5 @@
 import { DeploymentGraph } from "../../../dsl/DeploymentGraph";
-import { eliminate } from "../../../internal/graph/adjacencyList";
+import { eliminate } from "../../graph/adjacencyList";
 import { DeploymentGraphVertex } from "../../types/deploymentGraph";
 
 /**
